services/policymanagement-service/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysqlclient
python-dotenv
PyMuPDF
google-cloud-storage
passlib[bcrypt]
pyjwt